+++ examples/shop-with-blog/client/falcon-client.build.config.js
@@ -3,5 +3,6 @@ module.exports = {
   useWebmanifest: true,
   i18n: {
     resourcePackages: ['@deity/falcon-i18n']
-  }
+  },
+  moduleOverride: {}
 };