+++ packages/falcon-client/build-utils/webpack/config/create.js
@@ -10,6 +10,7 @@ const WebpackBar = require('webpackbar');
 const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 const VirtualModulesPlugin = require('webpack-virtual-modules');
 const FalconI18nLocalesPlugin = require('@deity/falcon-i18n-webpack-plugin');
+const NormalModuleOverridePlugin = require('@deity/normal-module-override-webpack-plugin');
 const errorOverlayMiddleware = require('react-dev-utils/errorOverlayMiddleware');
 const getCSSModuleLocalIdent = require('react-dev-utils/getCSSModuleLocalIdent');
 const LoadablePlugin = require('@loadable/webpack-plugin');
@@ -19,6 +20,15 @@ const { getClientEnv } = require('./env');
 const runPlugin = require('./runPlugin');
 
 const falconClientPolyfills = require.resolve('./../../polyfills');
+
+/**
+ * Create RegExp filter based on provided modules names
+ * @param {string[]} modules module names to filter on
+ * @returns {RegExp} RegExp filter
+ */
+function moduleFilter(modules) {
+  return new RegExp(`[\\\\/]node_modules[\\\\/](${modules.map(x => x.replace('/', '[\\\\/]')).join('|')})[\\\\/]`);
+}
 
 function getEsLintLoaderOptions(eslintRcPath, isDev) {
   const options = {
@@ -41,12 +51,12 @@ function getEsLintLoaderOptions(eslintRcPath, isDev) {
 function getBabelLoaderOptions(babelRcPath) {
   const options = {
     babelrc: true,
-    cacheDirectory: true,
     presets: []
   };
 
   const hasBabelRc = fs.existsSync(babelRcPath);
   if (!hasBabelRc) {
+    options.babelrc = false;
     options.presets.push(require.resolve('@deity/babel-preset-falcon-client'));
   }
 
@@ -100,31 +110,6 @@ function getStyleLoaders(target, env, cssLoaderOptions) {
   ];
 }
 
-function addVendorsBundle(modules = []) {
-  const moduleFilter = new RegExp(
-    `[\\\\/]node_modules[\\\\/](${modules.map(x => x.replace('/', '[\\\\/]')).join('|')})[\\\\/]`
-  );
-
-  return (config, { target, dev }) => {
-    if (target === 'web') {
-      config.output.filename = dev ? 'static/js/[name].js' : 'static/js/[name].[hash:8].js';
-
-      config.optimization = {
-        splitChunks: {
-          cacheGroups: {
-            vendor: {
-              name: 'vendors',
-              enforce: true,
-              chunks: 'initial',
-              test: moduleFilter
-            }
-          }
-        }
-      };
-    }
-  };
-}
-
 /**
  * Webpack configuration factory. It's the juice!
  * @param {'web' | 'node' } target target
@@ -134,7 +119,7 @@ function addVendorsBundle(modules = []) {
  */
 module.exports = (target = 'web', options, buildConfig) => {
   const { env, host, devServerPort, paths } = options;
-  const { useWebmanifest, plugins, modify, i18n } = buildConfig;
+  const { useWebmanifest, plugins, modify, i18n, moduleOverride } = buildConfig;
 
   // Define some useful shorthands.
   const IS_NODE = target === 'node';
@@ -348,6 +333,10 @@ module.exports = (target = 'web', options, buildConfig) => {
   }
 
   if (IS_WEB) {
+    config.entry = {
+      client: [falconClientPolyfills, require.resolve('pwacompat'), paths.ownClientIndexJs]
+    };
+    config.optimization = {};
     config.plugins = [
       new VirtualModulesPlugin({ [paths.ownWebmanifest]: '{}' }),
       new FalconI18nLocalesPlugin({
@@ -362,23 +351,11 @@ module.exports = (target = 'web', options, buildConfig) => {
         includeAllFileTypes: true,
         prettyPrint: true
       }),
-      new LoadablePlugin({
-        outputAsset: false,
-        writeToDisk: { filename: paths.appBuild }
-      })
+      new LoadablePlugin({ outputAsset: false, writeToDisk: { filename: paths.appBuild } })
     ];
 
     if (IS_DEV) {
-      // Setup Webpack Dev Server on port 3001 and
-      // specify our client entry point /client/index.js
-      config.entry = {
-        client: [
-          // We ship a few polyfills by default but only include them if React is being placed in the default path.
-          !clientEnv.raw.REACT_BUNDLE_PATH && falconClientPolyfills,
-          require.resolve('razzle-dev-utils/webpackHotDevClient'),
-          paths.ownClientIndexJs
-        ].filter(Boolean)
-      };
+      config.entry.client.push(require.resolve('razzle-dev-utils/webpackHotDevClient'));
 
       // Configure our client bundles output. Not the public path is to 3001.
       config.output = {
@@ -386,7 +363,7 @@ module.exports = (target = 'web', options, buildConfig) => {
         publicPath: `http://${host}:${devServerPort}/`,
         pathinfo: true,
         libraryTarget: 'var',
-        filename: 'static/js/bundle.js',
+        filename: 'static/js/[name].js',
         chunkFilename: 'static/js/[name].chunk.js',
         devtoolModuleFilenameTemplate: info => path.resolve(info.resourcePath).replace(/\\/g, '/')
       };
@@ -425,26 +402,14 @@ module.exports = (target = 'web', options, buildConfig) => {
         new webpack.HotModuleReplacementPlugin({ multiStep: true }),
         new webpack.DefinePlugin(clientEnv.stringified)
       ];
-
-      config.optimization = {};
     } else {
-      // Specify production entry point (/client/index.js)
-      config.entry = {
-        client: [
-          // We ship a few polyfills by default but only include them if React is being placed in the default path.
-          // If you are doing some vendor bundling, you'll need to require the @deity/falcon-client/build-utils/polyfills on your own.
-          !clientEnv.raw.REACT_BUNDLE_PATH && falconClientPolyfills,
-          paths.ownClientIndexJs
-        ].filter(Boolean)
-      };
-
       // Specify the client output directory and paths. Notice that we have
       // changed the publiPath to just '/' from http://localhost:3001. This is because
       // we will only be using one port in production.
       config.output = {
         path: paths.appBuildPublic,
         publicPath: options.publicPath,
-        filename: 'static/js/bundle.[chunkhash:8].js',
+        filename: 'static/js/[name].[hash:8].js',
         chunkFilename: 'static/js/[name].[chunkhash:8].chunk.js',
         libraryTarget: 'var'
       };
@@ -501,16 +466,49 @@ module.exports = (target = 'web', options, buildConfig) => {
             // Use multi-process parallel running to improve the build speed
             // Default number of concurrent runs: os.cpus().length - 1
             parallel: true,
-            cache: true, // Enable file caching
             sourceMap: !!devtool
           })
-        ]
+        ],
+        splitChunks: {
+          cacheGroups: {
+            polyfills: {
+              name: 'polyfills',
+              enforce: true,
+              priority: 100,
+              chunks: 'initial',
+              test: moduleFilter(['core-js', 'object-assign', 'whatwg-fetch', 'pwacompat'])
+            },
+            vendor: {
+              name: 'vendors',
+              enforce: true,
+              chunks: 'initial',
+              test: moduleFilter([
+                'apollo-cache-inmemory',
+                'apollo-client',
+                'apollo-link',
+                'apollo-link-http',
+                'apollo-utilities',
+                'i18next',
+                'i18next-xhr-backend',
+                'react',
+                'react-apollo',
+                'react-dom',
+                'react-google-tag-manager',
+                `react-helmet`,
+                'react-router',
+                'react-router-dom',
+                'history'
+              ])
+            }
+          }
+        }
       };
     }
   }
 
   config.plugins = [
     ...config.plugins,
+    new NormalModuleOverridePlugin(moduleOverride),
     new WebpackBar({
       minimal: options.isCI,
       color: colors.deityGreen,
@@ -518,27 +516,6 @@ module.exports = (target = 'web', options, buildConfig) => {
       compiledIn: true
     })
   ];
-
-  addVendorsBundle([
-    'apollo-cache-inmemory',
-    'apollo-client',
-    'apollo-link',
-    'apollo-link-http',
-    'apollo-link-state',
-    'apollo-utilities',
-    'node-fetch',
-    'i18next',
-    'i18next-xhr-backend',
-    '@deity/falcon-client/build-utils/polyfills',
-    'react',
-    'react-apollo',
-    'react-dom',
-    'react-google-tag-manager',
-    `react-helmet`,
-    'react-router',
-    'react-router-dom',
-    'history'
-  ])(config, { target, dev: IS_DEV });
 
   // Apply razzle plugins, if they are present in razzle.config.js
   if (Array.isArray(plugins)) {