+++ packages/falcon-client/src/index.js
@@ -1,14 +1,17 @@
 import http from 'http';
 import Logger from '@deity/falcon-logger';
-import { smartRequire } from '@loadable/server/lib/util';
 
-function falconWebServer() {
+function falconWebServer(port) {
   const { Server } = require('./server');
   const app = require('./clientApp');
   const bootstrap = require('./clientApp/bootstrap');
-  // eslint-disable-next-line
+  /* eslint-disable */
   const assetsManifest = require(process.env.ASSETS_MANIFEST);
-  const loadableStats = smartRequire(process.env.LOADABLE_STATS);
+  const loadableStats =
+    process.env.NODE_ENV === 'production'
+      ? require(process.env.LOADABLE_STATS)
+      : __non_webpack_require__(process.env.LOADABLE_STATS);
+  /* eslint-enable */
 
   /**
    * Creates an instance of Falcon web server
@@ -22,18 +25,19 @@ function falconWebServer() {
     webpackAssets: {
       webmanifest: assetsManifest[''].webmanifest
     },
+    port,
     loadableStats
   });
 }
 
-const server = falconWebServer();
+const port = parseInt(process.env.PORT, 10) || 3000;
+const server = falconWebServer(port);
 let currentWebServerHandler = server.callback();
 
 // Use `app#callback()` method here instead of directly
 // passing `app` as an argument to `createServer` (or use `app#listen()` instead)
 // @see https://github.com/koajs/koa/blob/master/docs/api/index.md#appcallback
 const httpServer = http.createServer(currentWebServerHandler);
-const port = parseInt(process.env.PORT, 10) || 3000;
 httpServer.listen(port, error => {
   if (error) {
     Logger.error(error);
@@ -50,7 +54,7 @@ if (module.hot) {
     Logger.log('🔁  HMR: Reloading server...');
 
     try {
-      const newHandler = falconWebServer().callback();
+      const newHandler = falconWebServer(port).callback();
       httpServer.removeListener('request', currentWebServerHandler);
       httpServer.on('request', newHandler);
       currentWebServerHandler = newHandler;