+++ packages/falcon-magento2-api/src/Magento2ApiBase.js
@@ -34,7 +34,7 @@ module.exports = class Magento2ApiBase extends ApiDataSource {
       const value = await this.cache.get({
         key: [this.name, this.session.storeCode || 'default', url].join(':'),
         callback: async () => {
-          const rawValue = await this.get(url);
+          const rawValue = await this.get(url, {}, { context: { useAdminToken: true } });
           return JSON.stringify(rawValue);
         },
         options: {
@@ -71,7 +71,6 @@ module.exports = class Magento2ApiBase extends ApiDataSource {
     }));
     this.storeConfigMap = _.keyBy(activeStoreConfigs, 'code');
     this.magentoConfig.locales = _.uniq(activeStoreConfigs.map(x => x.locale));
-    this.magentoConfig.defaultLocale = activeStoreConfigs.find(x => x.code === 'default').locale;
 
     activeStoreWebsites.forEach(storeWebsite => {
       const groups = [];