+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) are already released and available via npm. Versions without a date are not released yet.
 
+## Falcon next
+
+### Falcon Client next
+
+- added `.browserslistrc` support for javascript, fixed #293 ([#306](https://github.com/deity-io/falcon/pull/306))
+- added support for normal module replacement ([#328](https://github.com/deity-io/falcon/pull/328))
+- HTML5 autocomplete on checkout address form ([#330](https://github.com/deity-io/falcon/pull/330))
+- introduced `graphqlProxy` config flag to control GraphQL proxy via Falcon-Client route ([#337](https://github.com/deity-io/falcon/pull/337))
+
+### Normal Module Override Webpack Plugin v0.0.1
+
+- feat: first version of Normal Module Override Webpack Plugin ([#328](https://github.com/deity-io/falcon/pull/328))
+
+---
+
 ## Falcon 1.0 RC (2019-02-14)
 
 ### Falcon Shop Extension v0.3.0 (2019-02-14)
@@ -20,6 +35,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 - added `SearchProvider` with `SearchContext` that handles basic filtering ([#309](https://github.com/deity-io/falcon/pull/309))
 - added handling of dynamic menu ([#280](https://github.com/deity-io/falcon/pull/280))
+- updated `ApolloClient` to the newest version (v2.5.1) ([#333](https://github.com/deity-io/falcon/pull/333))
 
 ### Falcon Client v0.3.0 (2019-02-14)